requirements.txt
Django==4.0.4
psycopg2==2.9.5
djangorestframework==3.13.1
django-cors-headers==3.13.0
PyJWT==2.6.0